% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{css_theme}
\alias{css_theme}
\title{A minimal theme for CSS plots}
\usage{
css_theme(base_size = 11, base_family = "")
}
\arguments{
\item{base_size}{Base font size in points.}

\item{base_family}{Base font family.}
}
\value{
A \link[ggplot2:theme]{ggplot2::theme} object.
}
\description{
A minimal theme for CSS plots
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(wt, mpg)) + geom_point() + css_theme()
}
