% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cssig-package.R
\docType{package}
\name{cssig-package}
\alias{cssig}
\alias{cssig-package}
\title{cssig: Composite Selection Signals for Detecting Signatures of Selection}
\description{
Implements the composite selection signals (CSS) method of Randhawa et al. (2014) \doi{10.1186/1471-2156-15-34} and Randhawa et al. (2015) \doi{10.1534/g3.115.017772}, which unifies multiple constituent selection tests, such as the fixation index, cross-population extended haplotype homozygosity and the change in derived allele frequency, into a single non-parametric index based on the rank distribution of the constituent tests. Provides the core CSS statistic, sliding-window smoothing, empirical significance thresholds, genomic region calling, false discovery rate estimation, reciprocal (signed) cohort contrasts, helpers for computing the constituent statistics from allele frequencies, and a family of visualisations including Manhattan, mirrored Manhattan, per-chromosome, regional and circular plots.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://msk99.github.io/cssig/}
  \item \url{https://github.com/msk99/cssig}
  \item Report bugs at \url{https://github.com/msk99/cssig/issues}
}

}
\author{
\strong{Maintainer}: Mehar Singh Khatkar \email{mehar.khatkar@adelaide.edu.au}

Authors:
\itemize{
  \item Mehar Singh Khatkar \email{mehar.khatkar@adelaide.edu.au}
}

}
\keyword{internal}
