% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-diagnostics.R
\name{css_fdr_density}
\alias{css_fdr_density}
\title{Density of q-values inside called regions versus the rest of the genome}
\usage{
css_fdr_density(x, regions, fdr = 0.05, title = NULL, subtitle = NULL)
}
\arguments{
\item{x}{A `css_result` that has been through [css_fdr()].}

\item{regions}{A `css_regions` object.}

\item{fdr}{FDR level to mark with a vertical line. Default `0.05`.}

\item{title, subtitle}{Plot labels.}
}
\value{
A [ggplot2::ggplot] object.
}
\description{
Reproduces Figure 2 of Randhawa et al. (2014). A clear separation between the
two densities is the evidence that called regions carry a real excess of
selection signal rather than the tail of the genome-wide distribution.
}
\examples{
data(css_sim_small)
res <- css(css_input(css_sim_small,
                     tests = c(fst = "high", xpehh = "high", ddaf = "high")))
res <- css_fdr(res, method = "BH")
res <- css_threshold(css_smooth(res))
reg <- css_regions(res)
if (nrow(reg)) css_fdr_density(res, reg)

}
