% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-diagnostics.R
\name{css_pdist}
\alias{css_pdist}
\title{Histogram of CSS p-values}
\usage{
css_pdist(x, bins = 50L, title = NULL, subtitle = NULL)
}
\arguments{
\item{x}{A \code{css_result}.}

\item{bins}{Number of histogram bins. Default \code{50}.}

\item{title, subtitle}{Plot labels.}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} object.
}
\description{
Histogram of CSS p-values
}
\examples{
data(css_sim_small)
res <- css(css_input(css_sim_small,
                     tests = c(fst = "high", xpehh = "high", ddaf = "high")))
css_pdist(res)
}
