% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{css_sim_small}
\alias{css_sim_small}
\title{Two-chromosome subset of the simulated panel}
\format{
As \link{css_sim}, restricted to two chromosomes.
}
\usage{
css_sim_small
}
\description{
Chromosomes 1 and 2 of \link{css_sim}, carrying the incomplete and complete hard
sweeps respectively. Sized so that examples and tests run quickly.
}
\examples{
data(css_sim_small)
table(css_sim_small$chr)
}
\keyword{datasets}
